moving on to styling

+++ babel.rb
@@ -22,7 +22,7 @@ end
 #require application files
 require 'chain_evaluator.rb'
 require 'language_detector.rb'
-require 'deletion.rb'
+require 'admin_tasks.rb'
 require 'helper.rb'
 require 'language_module.rb'
 
@@ -57,7 +57,7 @@ class Chain
   #0 - L2attempt dispatched 
   #1 - L2attempt submitted
   #2 - L1attempt dispatched
-  #3 - L2 attempt submitted - chain completed.
+  #3 - L1attempt submitted - chain completed.
   property :progress, Integer, :default => 0
   property :score, Integer, :default => 0         #the score for the chain when completed 
   belongs_to :candidate
@@ -71,8 +71,8 @@ class L2attempt
   include DataMapper::Resource
   property :sentence, Text
   property :filled, Boolean, :default => false
-  property :recieved_at, DateTime
-  property :submitted_at, DateTime
+  property :recieved_at, EpochTime
+  property :submitted_at, EpochTime
   
   belongs_to :chain, :key => true
   belongs_to :player
@@ -83,8 +83,8 @@ class L1attempt
   include DataMapper::Resource
   property :sentence, Text
   property :filled, Boolean, :default => false
-  property :recieved_at, DateTime
-  property :submitted_at, DateTime
+  property :recieved_at, EpochTime
+  property :submitted_at, EpochTime
   
   belongs_to :chain, :key => true
   belongs_to :player
@@ -126,21 +126,26 @@ class Player
   def add_score score
     self.total_score += score
     self.save
-  end 
+  end
+  
+  def active_chains
+    activel2s = self.l2attempts.select{|l2| l2.chain.progress == 0}
+    activel1s = self.l1attempts.select{|l1| l1.chain.progress == 2}
+    activel1s.size + activel2s.size
+  end
 end
 
 class Scorecard
   include DataMapper::Resource
+  property :id, Serial
   property :report, Text
   property :viewed, Boolean, :default => false
+  property :created_at, DateTime
   
-  belongs_to :player, :key => true
-  belongs_to :chain, :key => true
+  belongs_to :player
+  belongs_to :chain
 end
   
-
-   
-
 DataMapper.finalize.auto_upgrade!
 
 enable :sessions
@@ -155,8 +160,13 @@ get '/' do
   erb :home
 end
 
+get '/about' do
+  @title = 'The Tower of Babel: How to play'
+  erb :about
+end
+
 get '/add_candidate' do    
-  @title = 'Home'
+  @title = 'Add Candidate'
   erb :add_candidate  
 end
 
@@ -165,8 +175,8 @@ post '/add_candidate' do
   @errors = []
   if params[:sentence].empty?
     @errors << "The sentence cannot be empty"
-  elsif params[:sentence].split.size < 3
-    @errors << "The sentence must contain at least three words"
+  elsif params[:sentence].split.size < 4
+    @errors << "The sentence must contain at least four words"
   end
   if params[:source] == params[:target]
     @errors << "The target language must be different from the source language"
@@ -194,20 +204,13 @@ post '/add_candidate' do
     else                          #if it is a guest, just save it
       s.save 
     end     
-    redirect '/'
+    flash[:notice] = "Your Candidate has been submitted to the tower"
+    redirect '/confirmation'
   end  
 end
 
-get'/admin_data' do
-  @candidates = Candidate.all
-  @chains = Chain.all
-  @players = Player.all
-  @title = 'Admin data'
-  erb :admin_data
-end
-
 #displays a users candidates and suggested translations
-get '/display' do
+get '/display_candidates' do
   unless @player
     @title = "Please log in"
     erb :not_logged_in
@@ -232,7 +235,8 @@ post '/register' do
   p.joined_at = Time.now
   if p.save
     session[:player] = p.id if p 
-    redirect '/'
+    flash[:notice] = "Your request to join the hordes of new Babylon has been submitted to the tower"
+    redirect '/confirmation'
   else
     @title = 'Registration Errors'
     @errors = p.errors
@@ -252,9 +256,28 @@ get '/login' do
   erb :login
 end
 
+
 post '/login' do
   player = Player.first(:name => params[:name], :password => params[:password])
-  session[:player] = player.id if player
+  
+  if player
+    session[:player] = player.id
+    unviewed_score_cards = Scorecard.all(:player => player, :viewed => false)
+    if unviewed_score_cards.size > 0
+      flash[:notice] = "You have new Score Cards waiting..."
+    end
+    @title = "Player Control#{player.name}"
+    @player = player
+    erb :player_control
+  else
+    flash[:notice] = "That username and password combination was not recognised by the system"
+    @title = "Username and password not recognised"
+    erb :login
+  end
+end
+
+get '/logout' do
+  session[:player] = nil
   redirect '/'
 end
 
@@ -263,28 +286,33 @@ get '/new_chain' do
     @title = "Please log in"
     erb :not_logged_in
   else
-    #Do not use candidates the player entered
-    candidates = Candidate.all(:player.not => @player)
-  
-    #Do not use candidates, on which the user previously initiated a chain
-    eligible_candidates = candidates.select do |can|
-                            inclusion = true
-                            can.chains.each do |ch|
-                              inclusion = false if ch.l2attempt.player == @player
-                            end
-                            inclusion
-                          end
-  
-    if (@candidate = eligible_candidates[rand(eligible_candidates.size)])
-      chain = Chain.create(:candidate => @candidate)
-      @l2attempt = L2attempt.create(:chain => chain, :player => @player, :recieved_at => Time.now)
-      chain.save
-      @l2attempt.save
-      @title = 'Your New Trans-mission'
-      erb :new_chain
-    else
-      @title = 'Lack of candidates in tower'
+    if @player.active_chains >= 5
+      @title = "Too many active chains"
       erb :apology
+    else
+      #Do not use candidates the player entered
+      candidates = Candidate.all(:player.not => @player)
+  
+      #Do not use candidates, on which the user previously initiated a chain
+      eligible_candidates = candidates.select do |can|
+                              inclusion = true
+                              can.chains.each do |ch|
+                                inclusion = false if ch.l2attempt.player == @player
+                              end
+                              inclusion
+                            end
+  
+      if (@candidate = eligible_candidates[rand(eligible_candidates.size)])
+        chain = Chain.create(:candidate => @candidate)
+        @l2attempt = L2attempt.create(:chain => chain, :player => @player, :recieved_at => Time.now)
+        chain.save
+        @l2attempt.save
+        @title = 'Your New Trans-mission'
+        erb :new_chain
+      else
+        @title = 'Lack of candidates in tower'
+        erb :apology
+      end
     end
   end
 end
@@ -293,34 +321,39 @@ post '/submit_l2' do
   chain = Chain.get(params[:chain].to_i)
   l2 = L2attempt.get(params[:chain].to_i)
   
-  sentence_inspector = LanguageDetector.new params[:sentence]
+  if l2.filled == true
+    @title = "Submission already recieved"
+    erb :already_recieved
+  else
+    sentence_inspector = LanguageDetector.new params[:sentence]
   
-  #if they are trying to cheat by using the same language
-  if sentence_inspector.language == chain.candidate.source && sentence_inspector.confidence > 0.4 
-    #penalise them and inform them of this
-    @title = 'Penalty'
-    @player = l2.player
-    @penalty = @player.total_score / 10
-    @player.total_score -= @penalty
-    @player.save 
-    l2.destroy
-    chain.destroy
-    erb :penalty
-  elsif sentence_inspector.language != chain.candidate.target
-    #don't penalise them but inform them the submission is unacceptable as language is not recognised target
-    @title = 'Your New Trans-mission'
-    @l2attempt = l2
-    @candidate = chain.candidate
-    flash[:notice] = "Your sentence was not recoginsed as #{chain.candidate.target_language}"
-    erb :new_chain
-  else  
-    l2.sentence = params[:sentence]
-    l2.filled = true
-    l2.submitted_at = Time.now
-    l2.chain.progress = 1         #stage L2 attempt filled
-    l2.save
-    flash[:notice] = "Your translation has been returned to the tower"
-    redirect '/confirmation'
+    #if they are trying to cheat by using the same language
+    if sentence_inspector.language == chain.candidate.source && sentence_inspector.confidence > 0.4 
+      #penalise them and inform them of this
+      @title = 'Penalty'
+      @player = l2.player
+      @penalty = @player.total_score / 10
+      @player.total_score -= @penalty
+      @player.save 
+      l2.destroy
+      chain.destroy
+      erb :penalty
+    elsif sentence_inspector.language != chain.candidate.target
+      #don't penalise them but inform them the submission is unacceptable as language is not recognised target
+      @title = 'Your New Trans-mission'
+      @l2attempt = l2
+      @candidate = chain.candidate
+      flash[:notice] = "Your sentence was not recognised as #{chain.candidate.target_language}"
+      erb :new_chain
+    else  
+      l2.sentence = params[:sentence]
+      l2.filled = true
+      l2.submitted_at = Time.now
+      l2.chain.progress = 1         #stage L2 attempt filled
+      l2.save
+      flash[:notice] = "Your translation has been returned to the tower"
+      redirect '/confirmation'
+    end
   end
 end
 
@@ -363,16 +396,28 @@ end
 
 post '/submit_l1' do
   l1 = L1attempt.get(params[:chain].to_i)
-  l1.sentence = params[:sentence]
   
-  #TO DO - check_language -- use google language detect
-  
-  l1.filled = true
-  l1.submitted_at = Time.now
-  l1.chain.progress = 3         #stage L1 attempt filled
-  l1.save
-  session[:chain] = params[:chain].to_i
-  redirect '/score_page'
+  if l1.filled == true
+    @title = "Submission already recieved"
+    erb :already_recieved
+  else 
+    sentence_inspector = LanguageDetector.new params[:sentence]
+    if sentence_inspector.language != l1.chain.candidate.source
+      @title = 'Complete Trans-mission'
+      @l1attempt = l1
+      @chain = l1.chain
+      flash[:notice] = "Your sentence was not recognised as #{l1.chain.candidate.source_language}"
+      erb :continue_chain
+    else
+      l1.sentence = params[:sentence]
+      l1.filled = true
+      l1.submitted_at = Time.now
+      l1.chain.progress = 3         #stage L1 attempt filled
+      l1.save
+      session[:chain] = params[:chain].to_i
+      redirect '/score_page'
+    end
+  end
 end
 
 get '/score_page' do
@@ -385,6 +430,52 @@ get '/score_page' do
   erb :score_page
 end
 
+get '/display_scorecards' do
+  unless @player
+    @title = "Please log in"
+    erb :not_logged_in
+  else
+    @unviewed_score_cards = Scorecard.all(:player => @player, :viewed => false, :order => :created_at.desc)
+    @viewed_score_cards = Scorecard.all(:player => @player, :viewed => true, :order => :created_at.desc)
+    @title = "My Score Cards"
+    erb :display_scorecards
+  end
+end
+
+get '/scorecard/:id' do
+  @score_card = Scorecard.get(params[:id])
+  @title = "Score Card #{params[:id]}"
+  erb :score_page
+end
+
+get '/display_chains' do
+  unless @player
+    @title = "Please log in"
+    erb :not_logged_in
+  else
+    l2attempts = L2attempt.all(:player => @player)
+    l1attempts = L1attempt.all(:player => @player)
+    @active_l2attempts = l2attempts.select{|l2| l2.chain.progress == 0}
+    @active_l1attempts = l1attempts.select{|l1| l1.chain.progress == 2}
+    @title = "My active chains"
+    erb :display_chains
+  end
+end
+
+get '/new_chain/:id' do
+  @l2attempt = L2attempt.get(params[:id])
+  @candidate = @l2attempt.chain.candidate
+  @title = "Your Trans-mission"
+  erb :new_chain
+end
+
+get '/continue_chain/:id' do
+  @chain = Chain.get(params[:id])
+  @l1attempt = @chain.l1attempt
+  @title = 'Complete this Trans-mission'
+  erb :continue_chain
+end
+
